Fix actions for related files list
Some improvements to related logs list
Created view of file from related files list
Fixed view in action button (class base type problem)
Changed upload file dialog to set type and category when file is uploaded instead of when the dialog is closed.  As soon as file is uploaded, it is shown in the file list.